+++ src/DynamicTreeDataGrid/Source/IDynamicTreeDataGridSource`1.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Controls;
 
-namespace DynamicTreeDataGrid;
+namespace DynamicTreeDataGrid.Source;
 
 public interface IDynamicTreeDataGridSource<TModel> : ITreeDataGridSource<TModel>, IDynamicTreeDataGridSource {
     // TODO: Add filter, state, maybe sort?
